+++ NEWS.md
@@ -1,4 +1,8 @@
-# gdalraster 1.11.0.9030 (dev)
+# gdalraster 1.11.0.9040 (dev)
+
+* add `apply_geotransform()`: convert raster column/row to geospatial x/y coordinates, wrapper of `GDALApplyGeoTransform()` in the GDAL API, operating on matrix input (the internal wrapper `.apply_geotransform()` is unchanged) (2024-06-11)
+
+* add `GDALRaster$apply_geotransform()`: class method alternative to calling the stand-alone function `apply_geotransform()` on an object of class `GDALRaster` (2024-06-11)
 
 * Doc: replace `paste0()` with `file.path()` in the examples (2024-06-10)